:recycle: refactoring(api): move LearningContentResourceNotFound
