+++ api/lib/infrastructure/datasources/learning-content/challenge-datasource.js
@@ -1,6 +1,6 @@
 import isEmpty from 'lodash/isEmpty.js';
 import * as datasource from './datasource.js';
-import { LearningContentResourceNotFound } from './LearningContentResourceNotFound.js';
+import { LearningContentResourceNotFound } from '../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
 
 const VALIDATED_CHALLENGE = 'validé';
 // donnée temporaire pour pix1d le temps d'arriver en « prod »
