+++ api/lib/infrastructure/datasources/learning-content/datasource.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { lcms } from '../../lcms.js';
-import { LearningContentResourceNotFound } from './LearningContentResourceNotFound.js';
+import { LearningContentResourceNotFound } from '../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
 import { LearningContentCache } from '../../caches/learning-content-cache.js';
 
 const _DatasourcePrototype = {
