+++ api/lib/infrastructure/repositories/course-repository.js
@@ -1,6 +1,6 @@
 import { Course } from '../../domain/models/Course.js';
 import { courseDatasource } from '../datasources/learning-content/course-datasource.js';
-import { LearningContentResourceNotFound } from '../datasources/learning-content/LearningContentResourceNotFound.js';
+import { LearningContentResourceNotFound } from '../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
 import { NotFoundError } from '../../domain/errors.js';
 
 function _toDomain(courseDataObject) {
