+++ api/tests/tooling/learning-content-builder/build-learning-content_test.js
@@ -1,7 +1,7 @@
 import { expect, learningContentBuilder, mockLearningContent } from '../../test-helper.js';
 import { areaDatasource } from '../../../lib/infrastructure/datasources/learning-content/area-datasource.js';
 import { competenceDatasource } from '../../../lib/infrastructure/datasources/learning-content/competence-datasource.js';
-import { thematicDatasource } from '../../../lib/infrastructure/datasources/learning-content/thematic-datasource.js';
+import { thematicDatasource } from '../../../src/shared/infrastructure/datasources/learning-content/thematic-datasource.js';
 import { tubeDatasource } from '../../../lib/infrastructure/datasources/learning-content/tube-datasource.js';
 import { skillDatasource } from '../../../lib/infrastructure/datasources/learning-content/skill-datasource.js';
 import { challengeDatasource } from '../../../lib/infrastructure/datasources/learning-content/challenge-datasource.js';
