+++ api/tests/unit/infrastructure/datasources/learning-content/challenge-datasource_test.js
@@ -3,7 +3,7 @@ import { expect, sinon, catchErr } from '../../../../test-helper.js';
 import { lcms } from '../../../../../lib/infrastructure/lcms.js';
 import { challengeDatasource } from '../../../../../lib/infrastructure/datasources/learning-content/challenge-datasource.js';
 import { learningContentCache } from '../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { LearningContentResourceNotFound } from '../../../../../lib/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
+import { LearningContentResourceNotFound } from '../../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | ChallengeDatasource', function () {
   let competence1,
